Refactor Syna into multiple subpackages (#16)
