+++ README.md
@@ -46,7 +46,7 @@ Visualize the computation graph for the fifth derivative of tanh(x) with respect
 ```python
 import syna
 import syna.functions as F
-from syna.utils import plot_dot_graph
+from syna import utils
 
 x = syna.tensor(1.0)
 y = F.tanh(x)
@@ -63,7 +63,7 @@ for i in range(iters):
 
 gx = x.grad
 gx.name = "gx" + str(iters + 1)
-plot_dot_graph(gx, verbose=False, to_file="tanh.svg")
+utils.viz.plot_dot_graph(gx, verbose=False, to_file="tanh.svg")
 ```
 
 The output graph is shown below.
@@ -76,7 +76,7 @@ Solve CartPole-v1 using the DQN algorithm.
 
 ```python
 from syna.algo.dqn import DQNAgent
-from syna.rl import Trainer
+from syna.utils.rl import Trainer
 
 trainer = Trainer(
     env_name="CartPole-v1",
