+++ docs/examples/dqn.rst
@@ -6,7 +6,7 @@ Solve CartPole-v1 using the DQN algorithm.
 .. code-block:: python
 
     from syna.algo.dqn import DQNAgent
-    from syna.rl import Trainer
+    from syna.utils.rl import Trainer
 
     trainer = Trainer(
         env_name="CartPole-v1",
@@ -15,4 +15,3 @@ Solve CartPole-v1 using the DQN algorithm.
             lr=3e-4,
         ),
     )
-    trainer.train()
