+++ examples/tanh.py
@@ -4,7 +4,7 @@ Need the dot binary from the graphviz package (www.graphviz.org).
 
 import syna
 import syna.functions as F
-from syna.utils import plot_dot_graph
+from syna import utils
 
 x = syna.tensor(1.0)
 y = F.tanh(x)
@@ -21,4 +21,4 @@ for i in range(iters):
 
 gx = x.grad
 gx.name = "gx" + str(iters + 1)
-plot_dot_graph(gx, verbose=False, to_file="tanh.svg")
+utils.viz.plot_dot_graph(gx, verbose=False, to_file="tanh.svg")
