+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "syna"
-version = "0.2.0"
+version = "0.3.0"
 description = "Syna is a lightweight machine learning framework inspired by DeZero."
 readme = "README.md"
 authors = [
