+++ src/syna/datasets/__init__.py
@@ -1,0 +1,1 @@
+from syna.datasets.datasets import Dataset
