+++ src/syna/models/__init__.py
@@ -1,0 +1,3 @@
+from syna.models.models import MLP, Model
+
+__all__ = ["Model", "MLP"]
